+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.4] - 2022-12-21
+
 ### Added
 
 - Add prometheus metrics about dex-app registrations on identity providers
@@ -49,7 +51,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added azure active directory provider.
 - Added initial implementation of the dex operator.
 
-[Unreleased]: https://github.com/giantswarm/dex-operator/compare/v0.1.3...HEAD
+[Unreleased]: https://github.com/giantswarm/dex-operator/compare/v0.1.4...HEAD
+[0.1.4]: https://github.com/giantswarm/dex-operator/compare/v0.1.3...v0.1.4
 [0.1.3]: https://github.com/giantswarm/dex-operator/compare/v0.1.2...v0.1.3
 [0.1.2]: https://github.com/giantswarm/dex-operator/compare/v0.1.1...v0.1.2
 [0.1.1]: https://github.com/giantswarm/dex-operator/compare/v0.1.0...v0.1.1